+++ source/ChanSort.Loader.SatcoDX/Channel.cs
@@ -1,12 +1,14 @@
 ﻿using System;
 using System.Globalization;
 using System.IO;
+using System.Text;
 using ChanSort.Api;
 
 namespace ChanSort.Loader.SatcoDX
 {
   internal class Channel : ChannelInfo
   {
+    private byte[] data;
     public int FileOffset { get; }
     public int Length { get; }
 
@@ -14,6 +16,7 @@ namespace ChanSort.Loader.SatcoDX
 
     internal Channel(int pos, string line, byte[] data, int start, int length, DvbStringDecoder decoder)
     {
+      this.data = data;
       this.FileOffset = start;
       this.Length = length;
       this.RecordIndex = pos;
@@ -41,14 +44,8 @@ namespace ChanSort.Loader.SatcoDX
       // 42: polarity
       this.Polarity = line[42] == '1' ? 'H' : 'V';
 
-      // 43-50 + (115-126 or 115-131): channel name
-      byte[] nameBytes = new byte[8+17];
-      var nameLen2 = Math.Min(length - 115, 17); // version 103 has 12 extra bytes for channel name, version 105 has 17
-      Array.Copy(data, start + 43, nameBytes, 0, 8);
-      Array.Copy(data, start + 115, nameBytes, 8, nameLen2);
-      decoder.GetChannelNames(nameBytes,0, nameBytes.Length, out var longName, out var shortName);
-      this.Name = longName.TrimEnd();
-      this.ShortName = shortName.TrimEnd();
+      // 43-50 + 115-126 in version 103 or 115-131 in version 105: channel name
+      this.ParseName(decoder);
 
       // 51-54: sat position
       var spos = line.Substring(51, 4).TrimStart('0');
@@ -82,6 +79,34 @@ namespace ChanSort.Loader.SatcoDX
       // 111-114: crypto code
     }
 
+    #endregion
+
+    #region ParseName()
+
+    /// <summary>
+    /// SATCODX103 files can contain channel names with unspecified implicit encoding, so we support reparsing based on a user selected default code page
+    /// </summary>
+    /// <param name="decoder"></param>
+    public void ParseName(DvbStringDecoder decoder)
+    {
+      var length = this.Length;
+      var start = this.FileOffset;
+
+      // 43-50 + 115-126 in version 103 or 115-131 in version 105: channel name
+      byte[] nameBytes = new byte[8 + 17];
+      var nameLen2 = Math.Min(length - 115, 17); // version 103 has 12 extra bytes for channel name, version 105 has 17
+      Array.Copy(data, start + 43, nameBytes, 0, 8);
+      Array.Copy(data, start + 115, nameBytes, 8, nameLen2);
+      
+      // I have seen format 103 files using only implicit CP1252 encoding for Umlauts, as well as format 105 with implicit UTF-8/explicit DVB-encoding
+      var oldDefaultEncoding = decoder.DefaultEncoding;
+      if (nameLen2 > 12)
+        decoder.DefaultEncoding = Encoding.UTF8;
+      decoder.GetChannelNames(nameBytes, 0, nameBytes.Length, out var longName, out var shortName);
+      decoder.DefaultEncoding = oldDefaultEncoding;
+      this.Name = longName.TrimEnd();
+      this.ShortName = shortName.TrimEnd();
+    }
     #endregion
   }
 }